requirements.txt
streamlit>=1.39.0
google-generativeai>=0.8.0
groq==0.4.1
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
pandas==2.1.3
SpeechRecognition==3.10.0

